refactor : schedule/modify 수정 로직 훅으로 분리
